[hotfix][network] Move RecordWriterDelegate to flink-streaming-java package
This interface is only used by StreamTask
