+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/io/RecordWriterDelegate.java
@@ -16,11 +16,12 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.io.network.api.writer;
+package org.apache.flink.streaming.runtime.io;
 
 import org.apache.flink.core.io.IOReadableWritable;
 import org.apache.flink.runtime.event.AbstractEvent;
 import org.apache.flink.runtime.io.AvailabilityProvider;
+import org.apache.flink.runtime.io.network.api.writer.RecordWriter;
 
 import java.io.IOException;
